takt: resolve-five-open-issues
